+++ docs/changelog.md
@@ -7,6 +7,7 @@
     * Added support for SCP style GitHub urls. Example: git@github.com:gamechanger/dusty.git
     * Added `image_requires_login` key to specs which can specify Docker images to pull. If set, Dusty will make sure the user is logged in to this image's repository before trying to pull the image
     * Optimized the `up` and `test` flows for some significant performance gains
+    * Dusty now shows the progress of image pulls on `dusty up`
 * **Misc**
     * Added support for Docker Compose 1.6
     * `dusty disk restore` now appends the `dusty-backup` suffix to the restore path if not provided, which should make the `restore` command a bit more intuitive to use